yolo_full/requirements.txt
# 深度学习框架
torch>=1.10.0
torchvision>=0.11.0
torchaudio>=0.10.0

# 计算机视觉
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
Pillow>=8.0.0

# 数据处理
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# GUI开发
PyQt5>=5.15.0
PyQt5-tools>=5.15.0

# YOLOv5依赖
PyYAML>=5.3.1
tqdm>=4.64.0
tensorboard>=2.4.1

# 图像标注工具
labelImg>=1.8.0

# 模型优化
onnx>=1.12.0
onnxruntime-gpu>=1.12.0  # GPU版本
# onnxruntime>=1.12.0  # CPU版本

# 工具库
requests>=2.26.0
scikit-learn>=0.24.0
scikit-image>=0.18.0

# 日志
loguru>=0.5.0

# 进度条
rich>=10.0.0

# 配置文件
python-dotenv>=0.19.0

# Jupyter支持
jupyter>=1.0.0
ipywidgets>=7.6.0
notebook>=6.4.0

# 测试
pytest>=6.2.0
pytest-cov>=2.12.0

# 代码质量
black>=21.7b0
flake8>=3.9.0
isort>=5.9.0

# 文档
Sphinx>=4.1.0
sphinx-rtd-theme>=0.5.2

# 其他工具
albumentations>=1.0.0  # 数据增强
imgaug>=0.4.0  # 数据增强
thop>=0.1.0  # 模型复杂度计算


